Allow front matter and pattern collections.
Adjust the matching code to allow collections to be defined in the front matter and by pattern.

Currently if you define collection in the front matter no patterns will match that file.

This aides in building something like a tagging system on a blog. All posts can be matched
with something like `'*.md'` or `'blog/**/*'` and then tags can be added in the individual
posts `collection` front matter.